Немножко потыкал => module export import, деконструкция сложная штука, было мало времени потыкать все.
