+++ index.html
@@ -59,8 +59,7 @@
 		<div class="wall-left"></div>
 		<div class="wall-right"></div>
 	</div>
-	<script src="logs.js"></script>
-	<script src="main.js"></script>
+	<script src="main.js" type="module"></script>
 </body>
 
 </html>
